Move frontend application code into app; modify paths to handle change
