+++ frontend/app/components/ToolCard.tsx
