+++ frontend/app/interfaces.ts
